replaces tree with proof and leaf
