+++ prover/circuits/invalidity_proof/bad_balance.go
@@ -2,18 +2,17 @@ package invalidity_proof
 
 import (
 	"github.com/consensys/gnark/frontend"
+	"github.com/consensys/linea-monorepo/prover/utils"
 )
 
 // BadBalanceCircuit defines the circuit for the transaction with insufficient balance.
 // This means account.Balance is less than Tx.Value + Tx.GasFee
 type BadBalanceCircuit struct {
-	// Transaction value
-	TxValue frontend.Variable
-	// Transaction GasFee
-	TxGasFee frontend.Variable
+	// Transaction cost = tx.Value + tx.Gas * tx.GasFeeCap
+	TxCost frontend.Variable
 	// sender address
-	TxFromAddress frontend.Variable
-	AccountTrie   AccountTrie
+	// TxFromAddress frontend.Variable
+	AccountTrie AccountTrie
 }
 
 // Define represent the constraints relevant to [BadBalanceCircuit]
@@ -22,15 +21,14 @@ func (circuit *BadBalanceCircuit) Define(api frontend.API) error {
 	var (
 		balance = circuit.AccountTrie.Account.Balance
 	)
-	// check that the Account.Balance < TxValue + TxGasFee + 1
-	api.AssertIsLessOrEqual(balance, api.Add(circuit.TxValue, circuit.TxGasFee, 1))
+	// check that the Account.Balance < TxCost + 1
+	api.AssertIsLessOrEqual(balance, api.Add(circuit.TxCost, 1))
 
 	circuit.AccountTrie.Define(api)
 
 	//@azam check that tx fields are related to  Tx.Hash  and then in the interconnection we show that
-	//FTx.Hash is included in the RollingHash
+	//FTx.Hash and FromAddress  = HKey is included in the RollingHash
 
-	//@Azam check that FromAddress is consistent with the account
 	return nil
 }
 
@@ -41,11 +39,17 @@ func (c *BadBalanceCircuit) Allocate(config Config) {
 
 // Assign the circuit from [AssigningInputs]
 func (c *BadBalanceCircuit) Assign(assi AssigningInputs) {
-
+	var (
+		txCost  = assi.Transaction.Cost()
+		balance = assi.AccountTrieInputs.Account.Balance
+	)
 	*c = BadBalanceCircuit{
-		TxValue:       assi.Transaction.Value(),
-		TxGasFee:      assi.Transaction.Cost(),
-		TxFromAddress: assi.FromAddress,
+		TxCost: txCost,
+	}
+
+	//sanity-check:
+	if txCost.Cmp(balance) != 1 {
+		utils.Panic("tried to generate a bad-balance proof for a valid transaction")
 	}
 
 	c.AccountTrie.Assign(assi.AccountTrieInputs)
