+++ prover/circuits/invalidity_proof/toy_test.go
@@ -1,0 +1,106 @@
+package invalidity_proof_test
+
+import (
+	"testing"
+
+	"github.com/consensys/gnark-crypto/ecc"
+	"github.com/consensys/gnark/frontend"
+	"github.com/consensys/gnark/frontend/cs/scs"
+	inval "github.com/consensys/linea-monorepo/prover/circuits/invalidity_proof"
+	ac "github.com/consensys/linea-monorepo/prover/crypto/state-management/accumulator"
+	"github.com/consensys/linea-monorepo/prover/crypto/state-management/hashtypes"
+	"github.com/consensys/linea-monorepo/prover/crypto/state-management/smt"
+	"github.com/consensys/linea-monorepo/prover/utils/types"
+	"github.com/consensys/linea-monorepo/prover/zkevm/prover/statemanager/accumulator"
+	"github.com/stretchr/testify/require"
+)
+
+// test [inval.BadNonceCircuit]
+func TestBadNonce(t *testing.T) {
+
+	// generate witness
+	config := &smt.Config{
+		HashFunc: hashtypes.MiMC,
+		Depth:    10,
+	}
+
+	assignment := genWitness(t, tcases, config)
+	witness, err := frontend.NewWitness(&assignment, ecc.BLS12_377.ScalarField())
+	require.NoError(t, err)
+
+	// allocate the circuit for merkleProof
+	var circuit inval.BadNonceCircuit
+	circuit.AccountTrie.MerkleProof.Proofs.Siblings = make([]frontend.Variable, config.Depth)
+
+	// compile the circuit
+	scs, err := frontend.Compile(
+		ecc.BLS12_377.ScalarField(),
+		scs.NewBuilder,
+		&circuit,
+	)
+	require.NoError(t, err)
+	// solve the circuit
+
+	err = scs.IsSolved(witness)
+	require.NoError(t, err)
+
+}
+
+// generates the witness for [inval.BadNonceCircuit]
+func genWitness(t *testing.T, tcases []TestCases, config *smt.Config) inval.BadNonceCircuit {
+
+	tree, proofs, leaves := genShomei(t, tcases, config)
+
+	// test just for one leaf
+	proof := proofs[1]
+	leaf := leaves[1]
+	root := tree.Root
+
+	//generate merkle tree witMerkle
+	var witMerkle inval.MerkleProofCircuit
+
+	witMerkle.Proofs.Siblings = make([]frontend.Variable, len(proof.Siblings))
+	for j := 0; j < len(proof.Siblings); j++ {
+		witMerkle.Proofs.Siblings[j] = proof.Siblings[j][:]
+	}
+	witMerkle.Proofs.Path = proof.Path
+	witMerkle.Leaf = leaf[:]
+
+	witMerkle.Root = root[:]
+
+	//generate witness for account and leafOpening
+	a := tcases[1].Account
+
+	account := types.GnarkAccount{
+		Nonce:    a.Nonce,
+		Balance:  a.Balance,
+		CodeSize: a.CodeSize,
+	}
+
+	account.StorageRoot = a.StorageRoot[:]
+	account.MimcCodeHash = a.MimcCodeHash[:]
+	account.KeccakCodeHashMSB = a.KeccakCodeHash[16:]
+	account.KeccakCodeHashLSB = a.KeccakCodeHash[:16]
+
+	hval := ac.Hash(config, a)
+
+	l := tcases[1].Leaf
+	leafOpening := accumulator.GnarkLeafOpening{
+		Prev: l.Prev,
+		Next: l.Next,
+	}
+
+	leafOpening.HKey = l.HKey[:]
+	leafOpening.HVal = hval[:]
+
+	var accountTrie = inval.AccountTrie{
+		MerkleProof: witMerkle,
+		LeafOpening: leafOpening,
+		Account:     account,
+	}
+
+	return inval.BadNonceCircuit{
+		TxNonce:     tcases[1].Tx.Nonce,
+		AccountTrie: accountTrie,
+	}
+}
